Remove unnecessary frame checking for tracing
After investigating, the logging framework automatically provides the depth required
using the depth option. Remove the currently unnecessary frame checking and
set the depth for the necessary and desired output.

Add line number.

Add a very simplistic test if the module is run as the main module. This assists
with debugging.